backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
bcrypt
python-dotenv
pandas
numpy
requests
beautifulsoup4
crawl4ai
httpx
openpyxl
fpdf2
python-multipart
